Unified the job files for IOS-XE, JUNOS and F5
