+++ Cave_of_Wonders/Cisco/IOS_XE/Show_Access_Lists/DevNet_Sandbox_CSR1000v_show_access_lists_mind_map.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<meta http-equiv="X-UA-Compatible" content="ie=edge">
+<title>Markmap</title>
+<style>
+* {
+  margin: 0;
+  padding: 0;
+}
+#mindmap {
+  display: block;
+  width: 100vw;
+  height: 100vh;
+}
+</style>
+<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/markmap-toolbar@0.1.6/dist/style.min.css">
+</head>
+<body>
+<svg id="mindmap"></svg>
+<script src="https://cdn.jsdelivr.net/npm/d3@6.6.0"></script><script src="https://cdn.jsdelivr.net/npm/markmap-view@0.2.6"></script><script src="https://cdn.jsdelivr.net/npm/markmap-toolbar@0.1.6/dist/index.umd.min.js"></script><script>(r => {
+          setTimeout(r);
+        })(() => {
+  const {
+    markmap,
+    mm
+  } = window;
+  const toolbar = new markmap.Toolbar();
+  toolbar.attach(mm);
+  const el = toolbar.render();
+  el.setAttribute('style', 'position:absolute;bottom:20px;right:20px');
+  document.body.append(el);
+})</script><script>((getMarkmap, getOptions, data) => {
+        const {
+          Markmap
+        } = getMarkmap();
+        window.mm = Markmap.create('svg#mindmap', getOptions == null ? void 0 : getOptions(), data);
+      })(() => window.markmap,null,{"t":"heading","d":1,"p":{"lines":[1,2]},"v":"Show IP Access-Lists","c":[{"t":"thead","d":3,"p":{"lines":[2,3]},"v":"","c":[{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Access Control List"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Access Control Entry"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Permission"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Logging"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Source Network"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Destination Network"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"L3 Protocol"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"L4 Protocol"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Operator"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Port"}]},{"t":"tbody","d":3,"p":{"lines":[4,10]},"v":"","c":[{"t":"tr","d":4,"p":{},"v":"","c":[{"t":"td","d":5,"p":{},"v":"IPvZero"},{"t":"td","d":5,"p":{},"v":"10"},{"t":"td","d":5,"p":{},"v":"permit"},{"t":"td","d":5,"p":{},"v":"log-none"},{"t":"td","d":5,"p":{},"v":"any"},{"t":"td","d":5,"p":{},"v":"any any"},{"t":"td","d":5,"p":{},"v":"icmp,N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"}]},{"t":"tr","d":4,"p":{},"v":"","c":[{"t":"td","d":5,"p":{},"v":"IPvZero"},{"t":"td","d":5,"p":{},"v":"20"},{"t":"td","d":5,"p":{},"v":"deny"},{"t":"td","d":5,"p":{},"v":"log-none"},{"t":"td","d":5,"p":{},"v":"any"},{"t":"td","d":5,"p":{},"v":"host 8.8.8.8 host 8.8.8.8"},{"t":"td","d":5,"p":{},"v":"tcp,N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"}]},{"t":"tr","d":4,"p":{},"v":"","c":[{"t":"td","d":5,"p":{},"v":"IPvZero"},{"t":"td","d":5,"p":{},"v":"30"},{"t":"td","d":5,"p":{},"v":"permit"},{"t":"td","d":5,"p":{},"v":"log-none"},{"t":"td","d":5,"p":{},"v":"any"},{"t":"td","d":5,"p":{},"v":"any any"},{"t":"td","d":5,"p":{},"v":"ipv4,N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"}]},{"t":"tr","d":4,"p":{},"v":"","c":[{"t":"td","d":5,"p":{},"v":"meraki-fqdn-dns"},{"t":"td","d":5,"p":{},"v":"No ACEs"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"}]},{"t":"tr","d":4,"p":{},"v":"","c":[{"t":"td","d":5,"p":{},"v":"IPvZero"},{"t":"td","d":5,"p":{},"v":"20"},{"t":"td","d":5,"p":{},"v":"deny"},{"t":"td","d":5,"p":{},"v":"log-none,N/A,Destination Protocol,N/A,TCP,eq"},{"t":"td","d":5,"p":{},"v":"22"}]},{"t":"tr","d":4,"p":{},"v":"","c":[{"t":"td","d":5,"p":{},"v":"meraki-fqdn-dns"},{"t":"td","d":5,"p":{},"v":"No ACEs"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"},{"t":"td","d":5,"p":{},"v":"N/A"}]}]}]})</script>
+</body>
+</html>
