+++ Cave_of_Wonders/Cisco/IOS_XE/Show_Etherchannel_Summary/DevNet_Sandbox_CSR1000v_show_etherchannel_summary.json
@@ -1,0 +1,4 @@
+{
+    "number_of_aggregators": 0,
+    "number_of_lag_in_use": 0
+}
