+++ Cave_of_Wonders/Cisco/IOS_XE/Show_Etherchannel_Summary/DevNet_Sandbox_CSR1000v_show_etherchannel_summary.yaml
@@ -1,0 +1,4 @@
+!!python/object/new:genie.conf.base.utils.QDict
+dictitems:
+  number_of_aggregators: 0
+  number_of_lag_in_use: 0
