+++ Cave_of_Wonders/Cisco/IOS_XE/Show_Etherchannel_Summary/DevNet_Sandbox_CSR1000v_show_etherchannel_summary_totals.csv
@@ -1,0 +1,2 @@
+Number of Aggregators,Number of LAG in use
+0,0
