+++ Cave_of_Wonders/Cisco/IOS_XE/Show_Etherchannel_Summary/DevNet_Sandbox_CSR1000v_show_etherchannel_summary_totals.html
@@ -1,0 +1,14 @@
+
+<html><body>
+<h1>Show Etherchannel Summary</h1>
+<table style="width:100%">
+  <tr>
+    <th>Number of Aggregators</th>
+    <th>Number of LAG In Use</th>
+  </tr>
+  <tr>
+    <td>0</td>
+    <td>0</td>
+  </tr>
+</table>
+</body></html>
