+++ Cave_of_Wonders/Cisco/IOS_XE/Show_Etherchannel_Summary/DevNet_Sandbox_CSR1000v_show_etherchannel_summary_totals.md
@@ -1,0 +1,5 @@
+
+# Etherchannel Totals
+| Number of Aggregators | Number of LAG in use |
+| --------------------- | -------------------- |
+| 0 | 0 |
