+++ Cave_of_Wonders/Cisco/IOS_XE/Show_IP_ARP/DevNet_Sandbox_CSR1000v_show_ip_arp.csv
@@ -1,0 +1,4 @@
+Interface,Neighbor IP Address,MAC Address,Type,Origin,Age,Protocol
+GigabitEthernet1,10.10.20.28,0050.56bf.490f,ARPA,dynamic,0,Internet
+GigabitEthernet1,10.10.20.48,0050.56bf.78ac,ARPA,static,-,Internet
+GigabitEthernet1,10.10.20.254,0050.56bf.d636,ARPA,dynamic,89,Internet
