+++ Cave_of_Wonders/Cisco/IOS_XE/Show_IP_ARP/DevNet_Sandbox_CSR1000v_show_ip_arp.html
@@ -1,0 +1,42 @@
+
+<html><body>
+<h1>Show IP ARP (Global Routing Table)</h1>
+<table style="width:100%">
+  <tr>
+    <th>Interface</th>
+    <th>Neighbor IP Address</th>
+    <th>MAC Address</th>
+    <th>Type</th>
+    <th>Origin</th>
+    <th>Age</th>
+    <th>Protocol</th>
+  </tr>
+  <tr>
+    <td>GigabitEthernet1</td>
+    <td>10.10.20.28</td>
+    <td>0050.56bf.490f</td>
+    <td>ARPA</td>
+    <td>dynamic</td>
+    <td>0</td>
+    <td>Internet</td>
+  </tr>
+  <tr>
+    <td>GigabitEthernet1</td>
+    <td>10.10.20.48</td>
+    <td>0050.56bf.78ac</td>
+    <td>ARPA</td>
+    <td>static</td>
+    <td>-</td>
+    <td>Internet</td>
+  </tr>
+  <tr>
+    <td>GigabitEthernet1</td>
+    <td>10.10.20.254</td>
+    <td>0050.56bf.d636</td>
+    <td>ARPA</td>
+    <td>dynamic</td>
+    <td>89</td>
+    <td>Internet</td>
+  </tr>
+</table>
+</body></html>
