+++ Cave_of_Wonders/Cisco/IOS_XE/Show_IP_ARP/DevNet_Sandbox_CSR1000v_show_ip_arp.yaml
@@ -1,0 +1,27 @@
+!!python/object/new:genie.conf.base.utils.QDict
+dictitems:
+  interfaces:
+    GigabitEthernet1:
+      ipv4:
+        neighbors:
+          10.10.20.254:
+            age: '89'
+            ip: 10.10.20.254
+            link_layer_address: 0050.56bf.d636
+            origin: dynamic
+            protocol: Internet
+            type: ARPA
+          10.10.20.28:
+            age: '0'
+            ip: 10.10.20.28
+            link_layer_address: 0050.56bf.490f
+            origin: dynamic
+            protocol: Internet
+            type: ARPA
+          10.10.20.48:
+            age: '-'
+            ip: 10.10.20.48
+            link_layer_address: 0050.56bf.78ac
+            origin: static
+            protocol: Internet
+            type: ARPA
