+++ Cave_of_Wonders/Cisco/IOS_XE/Show_IP_Interface_Brief/DevNet_Sandbox_CSR1000v_show_ip_int_brief.html
@@ -1,0 +1,38 @@
+
+<html><body>
+<h1>Show IP Interface Brief</h1>
+<table style="width:100%">
+  <tr>
+    <th>Interface</th>
+    <th>IP Address</th>
+    <th>Status</th>
+    <th>Protocol</th>
+    <th>Method</th>   
+    <th>Interface is OK</th>   
+  </tr>  
+  <tr>
+    <td>GigabitEthernet1</td>
+    <td>10.10.20.48</td>
+    <td>up</td>
+    <td>up</td>
+    <td>NVRAM</td>
+    <td>YES</td>
+  </tr>  
+  <tr>
+    <td>GigabitEthernet2</td>
+    <td>unassigned</td>
+    <td>administratively down</td>
+    <td>down</td>
+    <td>NVRAM</td>
+    <td>YES</td>
+  </tr>  
+  <tr>
+    <td>GigabitEthernet3</td>
+    <td>unassigned</td>
+    <td>administratively down</td>
+    <td>down</td>
+    <td>NVRAM</td>
+    <td>YES</td>
+  </tr>
+</table>
+</body></html>
