+++ Cave_of_Wonders/Cisco/IOS_XE/Show_IP_Interface_Brief/DevNet_Sandbox_CSR1000v_show_ip_int_brief.json
@@ -1,0 +1,25 @@
+{
+    "interface": {
+        "GigabitEthernet1": {
+            "interface_is_ok": "YES",
+            "ip_address": "10.10.20.48",
+            "method": "NVRAM",
+            "protocol": "up",
+            "status": "up"
+        },
+        "GigabitEthernet2": {
+            "interface_is_ok": "YES",
+            "ip_address": "unassigned",
+            "method": "NVRAM",
+            "protocol": "down",
+            "status": "administratively down"
+        },
+        "GigabitEthernet3": {
+            "interface_is_ok": "YES",
+            "ip_address": "unassigned",
+            "method": "NVRAM",
+            "protocol": "down",
+            "status": "administratively down"
+        }
+    }
+}
