+++ Cave_of_Wonders/Cisco/IOS_XE/Show_IP_Interface_Brief/DevNet_Sandbox_CSR1000v_show_ip_int_brief.md
@@ -1,0 +1,7 @@
+
+# Show IP Interface Brief
+| Interface | IP Address | Status | Protocol | Method | Interface is OK |
+| --------- | ---------- | ------ | -------- | ------ | --------------- |
+| GigabitEthernet1 | 10.10.20.48 | up | up | NVRAM | YES |
+| GigabitEthernet2 | unassigned | administratively down | down | NVRAM | YES |
+| GigabitEthernet3 | unassigned | administratively down | down | NVRAM | YES |
