+++ Cave_of_Wonders/Cisco/IOS_XE/Show_IP_Interface_Brief/DevNet_Sandbox_CSR1000v_show_ip_int_brief_mind_map.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<meta http-equiv="X-UA-Compatible" content="ie=edge">
+<title>Markmap</title>
+<style>
+* {
+  margin: 0;
+  padding: 0;
+}
+#mindmap {
+  display: block;
+  width: 100vw;
+  height: 100vh;
+}
+</style>
+<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/markmap-toolbar@0.1.6/dist/style.min.css">
+</head>
+<body>
+<svg id="mindmap"></svg>
+<script src="https://cdn.jsdelivr.net/npm/d3@6.6.0"></script><script src="https://cdn.jsdelivr.net/npm/markmap-view@0.2.6"></script><script src="https://cdn.jsdelivr.net/npm/markmap-toolbar@0.1.6/dist/index.umd.min.js"></script><script>(r => {
+          setTimeout(r);
+        })(() => {
+  const {
+    markmap,
+    mm
+  } = window;
+  const toolbar = new markmap.Toolbar();
+  toolbar.attach(mm);
+  const el = toolbar.render();
+  el.setAttribute('style', 'position:absolute;bottom:20px;right:20px');
+  document.body.append(el);
+})</script><script>((getMarkmap, getOptions, data) => {
+        const {
+          Markmap
+        } = getMarkmap();
+        window.mm = Markmap.create('svg#mindmap', getOptions == null ? void 0 : getOptions(), data);
+      })(() => window.markmap,null,{"t":"heading","d":1,"p":{"lines":[1,2]},"v":"Show IP Interface Brief","c":[{"t":"thead","d":3,"p":{"lines":[2,3]},"v":"","c":[{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Interface"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"IP Address"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Status"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Protocol"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Method"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Interface is OK"}]},{"t":"tbody","d":3,"p":{"lines":[4,7]},"v":"","c":[{"t":"tr","d":4,"p":{},"v":"","c":[{"t":"td","d":5,"p":{},"v":"GigabitEthernet1"},{"t":"td","d":5,"p":{},"v":"10.10.20.48"},{"t":"td","d":5,"p":{},"v":"up"},{"t":"td","d":5,"p":{},"v":"up"},{"t":"td","d":5,"p":{},"v":"NVRAM"},{"t":"td","d":5,"p":{},"v":"YES"}]},{"t":"tr","d":4,"p":{},"v":"","c":[{"t":"td","d":5,"p":{},"v":"GigabitEthernet2"},{"t":"td","d":5,"p":{},"v":"unassigned"},{"t":"td","d":5,"p":{},"v":"administratively down"},{"t":"td","d":5,"p":{},"v":"down"},{"t":"td","d":5,"p":{},"v":"NVRAM"},{"t":"td","d":5,"p":{},"v":"YES"}]},{"t":"tr","d":4,"p":{},"v":"","c":[{"t":"td","d":5,"p":{},"v":"GigabitEthernet3"},{"t":"td","d":5,"p":{},"v":"unassigned"},{"t":"td","d":5,"p":{},"v":"administratively down"},{"t":"td","d":5,"p":{},"v":"down"},{"t":"td","d":5,"p":{},"v":"NVRAM"},{"t":"td","d":5,"p":{},"v":"YES"}]}]}]})</script>
+</body>
+</html>
