+++ Cave_of_Wonders/Cisco/IOS_XE/Show_IP_Route/DevNet_Sandbox_CSR1000v_show_ip_route.html
@@ -1,0 +1,34 @@
+
+<html><body>
+<h1>Show IP Route</h1>
+<table style="width:100%">
+  <tr>
+    <th>VRF</th>
+    <th>Address Family</th>
+    <th>Route</th>
+    <th>Active</th>
+    <th>Metric</th>
+    <th>Route Preference</th>   
+    <th>Source Protocol</th>
+    <th>Source Protocol Code</th>
+    <th>Next Hop Number</th>
+    <th>Outgoing Interface</th>
+    <th>Updated</th>
+  </tr>
+  <tr>
+    <td>default</td>
+    <td>ipv4</td>
+    <td>0.0.0.0/0</td>
+    <td>True</td>
+    <td>0</td>   
+    <td>1</td>
+    <td>static</td>
+    <td>S*</td>
+    <td>1</td>
+    <td>10.10.20.254</td>
+    <td>GigabitEthernet1</td>
+    <td>GigabitEthernet1</td>
+    <td></td>
+  </tr>
+</table>
+</body></html>
