+++ Cave_of_Wonders/Cisco/IOS_XE/Show_IP_Route/DevNet_Sandbox_CSR1000v_show_ip_route.json
@@ -1,0 +1,55 @@
+{
+    "vrf": {
+        "default": {
+            "address_family": {
+                "ipv4": {
+                    "routes": {
+                        "0.0.0.0/0": {
+                            "active": true,
+                            "metric": 0,
+                            "next_hop": {
+                                "next_hop_list": {
+                                    "1": {
+                                        "index": 1,
+                                        "next_hop": "10.10.20.254",
+                                        "outgoing_interface": "GigabitEthernet1"
+                                    }
+                                }
+                            },
+                            "route": "0.0.0.0/0",
+                            "route_preference": 1,
+                            "source_protocol": "static",
+                            "source_protocol_codes": "S*"
+                        },
+                        "10.10.20.0/24": {
+                            "active": true,
+                            "next_hop": {
+                                "outgoing_interface": {
+                                    "GigabitEthernet1": {
+                                        "outgoing_interface": "GigabitEthernet1"
+                                    }
+                                }
+                            },
+                            "route": "10.10.20.0/24",
+                            "source_protocol": "connected",
+                            "source_protocol_codes": "C"
+                        },
+                        "10.10.20.48/32": {
+                            "active": true,
+                            "next_hop": {
+                                "outgoing_interface": {
+                                    "GigabitEthernet1": {
+                                        "outgoing_interface": "GigabitEthernet1"
+                                    }
+                                }
+                            },
+                            "route": "10.10.20.48/32",
+                            "source_protocol": "local",
+                            "source_protocol_codes": "L"
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
