+++ Cave_of_Wonders/Cisco/IOS_XE/Show_IP_Route/DevNet_Sandbox_CSR1000v_show_ip_route.md
@@ -1,0 +1,5 @@
+
+# Show IP Route
+| VRF | Address Family | Route | Active | Metric | Route Preference | Source Protocol | Source Protocol Code | Next Hop Number | Next Hop | Outgoing Interface | Updated |
+| --- | -------------- | ----- | ------ | ------ | ---------------- | --------------- | -------------------- | --------------- | -------- | ------------------ | ------- |
+| default | ipv4 | 0.0.0.0/0 | True | 0 | 1 | static | S* | 1 | 10.10.20.254 | GigabitEthernet1 |  |
