+++ Cave_of_Wonders/Cisco/IOS_XE/Show_IP_Route/DevNet_Sandbox_CSR1000v_show_ip_route_netgraph.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <link href="/src/netjsongraph.css" rel="stylesheet">
+        <link href="/src/netjsongraph-theme.css" rel="stylesheet">
+    </head>
+    <body>
+        <script src="/lib/d3.min.js"></script>
+        <script src="/src/netjsongraph.js"></script>
+        <script>d3.netJsonGraph("DevNet_Sandbox_CSR1000v_show_ip_route_netgraph.json");</script>
+    </body>
+</html>
