+++ Cave_of_Wonders/Cisco/IOS_XE/Show_Interfaces/DevNet_Sandbox_CSR1000v_show_int.html
@@ -1,0 +1,122 @@
+
+<html><body>
+<h1>Show Interfaces</h1>
+<table style="width:100%">
+  <tr>
+    <th>Interface</th>
+    <th>Description</th>
+    <th>Type</th>
+    <th>Connected</th>
+    <th>Enabled</th>
+    <th>Bandwidth</th>
+    <th>Port Speed</th>
+    <th>Port Duplex</th>
+    <th>IPv4 Address</th>
+    <th>Line Protocol</th>
+    <th>Operational Status</th>
+    <th>MAC Address</th>
+    <th>Media Type</th>
+    <th>Delay</th>
+    <th>Encapsulation</th>
+    <th>Keepalive</th>
+    <th>MTU</th>
+    <th>Port Channel Member</th>
+    <th>Port Channel Member Interface</th>
+    <th>Input CRC Errors</th>
+    <th>Input Errors</th>
+    <th>Output Errors</th>
+    <th>Input Packets</th>
+    <th>Output Packets</th>
+    <th>QoS Strategy</th>
+    <th>Input Queue Drops</th>
+    <th>Total Output Drops</th>
+    </tr>
+  <tr>
+    <td>GigabitEthernet1</td>
+    <td>MANAGEMENT INTERFACE - DON'T TOUCH ME</td>
+    <td>CSR vNIC</td>
+    <td>Not Connected</td>
+    <td>True</td>
+    <td>1000000</td>
+    <td>1000mbps</td>
+    <td>full</td>
+    <td>No IP Address</td>
+    <td>up</td>
+    <td>up</td>
+    <td>0050.56bf.78ac</td>
+    <td>Virtual</td>
+    <td>10</td>
+    <td>arpa</td>
+    <td>10</td>
+    <td>1500</td>
+    <td>False</td>
+    <td>Stand Alone Port</td>
+    <td>0</td>
+    <td>0</td>
+    <td>0</td>
+    <td>7335</td>
+    <td>8317</td>
+    <td>fifo</td>
+    <td>0</td>
+    <td>0</td>
+  </tr>
+  <tr>
+    <td>GigabitEthernet2</td>
+    <td>Network Interface</td>
+    <td>CSR vNIC</td>
+    <td>Not Connected</td>
+    <td>False</td>
+    <td>1000000</td>
+    <td>1000mbps</td>
+    <td>full</td>
+    <td>No IP Address</td>
+    <td>down</td>
+    <td>down</td>
+    <td>0050.56bf.4ea3</td>
+    <td>Virtual</td>
+    <td>10</td>
+    <td>arpa</td>
+    <td>10</td>
+    <td>1500</td>
+    <td>False</td>
+    <td>Stand Alone Port</td>
+    <td>0</td>
+    <td>0</td>
+    <td>0</td>
+    <td>5</td>
+    <td>0</td>
+    <td>fifo</td>
+    <td>0</td>
+    <td>0</td>
+  </tr>
+  <tr>
+    <td>GigabitEthernet3</td>
+    <td>Network Interface</td>
+    <td>CSR vNIC</td>
+    <td>Not Connected</td>
+    <td>False</td>
+    <td>1000000</td>
+    <td>1000mbps</td>
+    <td>full</td>
+    <td>No IP Address</td>
+    <td>down</td>
+    <td>down</td>
+    <td>0050.56bf.7db4</td>
+    <td>Virtual</td>
+    <td>10</td>
+    <td>arpa</td>
+    <td>10</td>
+    <td>1500</td>
+    <td>False</td>
+    <td>Stand Alone Port</td>
+    <td>0</td>
+    <td>0</td>
+    <td>0</td>
+    <td>0</td>
+    <td>0</td>
+    <td>fifo</td>
+    <td>0</td>
+    <td>0</td>
+  </tr>
+</table>
+</body></html>
