+++ Cave_of_Wonders/Cisco/IOS_XE/Show_Interfaces/DevNet_Sandbox_CSR1000v_show_int.yaml
@@ -1,0 +1,243 @@
+!!python/object/new:genie.conf.base.utils.QDict
+dictitems:
+  GigabitEthernet1:
+    arp_timeout: 04:00:00
+    arp_type: arpa
+    auto_negotiate: true
+    bandwidth: 1000000
+    counters:
+      in_broadcast_pkts: 0
+      in_crc_errors: 0
+      in_errors: 0
+      in_frame: 0
+      in_giants: 0
+      in_ignored: 0
+      in_mac_pause_frames: 0
+      in_multicast_pkts: 0
+      in_no_buffer: 0
+      in_octets: 797733
+      in_overrun: 0
+      in_pkts: 7335
+      in_runts: 0
+      in_throttles: 0
+      in_watchdog: 0
+      last_clear: never
+      out_babble: 0
+      out_buffer_failure: 0
+      out_buffers_swapped: 0
+      out_collision: 0
+      out_deferred: 0
+      out_errors: 0
+      out_interface_resets: 0
+      out_late_collision: 0
+      out_lost_carrier: 0
+      out_mac_pause_frames: 0
+      out_no_carrier: 0
+      out_octets: 2366847
+      out_pkts: 8317
+      out_underruns: 0
+      out_unknown_protocl_drops: 0
+      rate:
+        in_rate: 3000
+        in_rate_pkts: 3
+        load_interval: 300
+        out_rate: 3000
+        out_rate_pkts: 3
+    delay: 10
+    description: MANAGEMENT INTERFACE - DON'T TOUCH ME
+    duplex_mode: full
+    enabled: true
+    encapsulations:
+      encapsulation: arpa
+    flow_control:
+      receive: false
+      send: false
+    ipv4:
+      10.10.20.48/24:
+        ip: 10.10.20.48
+        prefix_length: '24'
+    keepalive: 10
+    last_input: 00:00:00
+    last_output: 00:00:00
+    line_protocol: up
+    link_type: auto
+    mac_address: 0050.56bf.78ac
+    media_type: Virtual
+    mtu: 1500
+    oper_status: up
+    output_hang: never
+    phys_address: 0050.56bf.78ac
+    port_channel:
+      port_channel_member: false
+    port_speed: 1000mbps
+    queues:
+      input_queue_drops: 0
+      input_queue_flushes: 0
+      input_queue_max: 375
+      input_queue_size: 1
+      output_queue_max: 40
+      output_queue_size: 0
+      queue_strategy: fifo
+      total_output_drop: 0
+    reliability: 255/255
+    rxload: 1/255
+    txload: 1/255
+    type: CSR vNIC
+  GigabitEthernet2:
+    arp_timeout: 04:00:00
+    arp_type: arpa
+    auto_negotiate: true
+    bandwidth: 1000000
+    counters:
+      in_broadcast_pkts: 0
+      in_crc_errors: 0
+      in_errors: 0
+      in_frame: 0
+      in_giants: 0
+      in_ignored: 0
+      in_mac_pause_frames: 0
+      in_multicast_pkts: 0
+      in_no_buffer: 0
+      in_octets: 300
+      in_overrun: 0
+      in_pkts: 5
+      in_runts: 0
+      in_throttles: 0
+      in_watchdog: 0
+      last_clear: never
+      out_babble: 0
+      out_buffer_failure: 0
+      out_buffers_swapped: 0
+      out_collision: 0
+      out_deferred: 0
+      out_errors: 0
+      out_interface_resets: 0
+      out_late_collision: 0
+      out_lost_carrier: 1
+      out_mac_pause_frames: 0
+      out_no_carrier: 0
+      out_octets: 0
+      out_pkts: 0
+      out_underruns: 0
+      out_unknown_protocl_drops: 0
+      rate:
+        in_rate: 0
+        in_rate_pkts: 0
+        load_interval: 300
+        out_rate: 0
+        out_rate_pkts: 0
+    delay: 10
+    description: Network Interface
+    duplex_mode: full
+    enabled: false
+    encapsulations:
+      encapsulation: arpa
+    flow_control:
+      receive: false
+      send: false
+    keepalive: 10
+    last_input: never
+    last_output: never
+    line_protocol: down
+    link_type: auto
+    mac_address: 0050.56bf.4ea3
+    media_type: Virtual
+    mtu: 1500
+    oper_status: down
+    output_hang: never
+    phys_address: 0050.56bf.4ea3
+    port_channel:
+      port_channel_member: false
+    port_speed: 1000mbps
+    queues:
+      input_queue_drops: 0
+      input_queue_flushes: 0
+      input_queue_max: 375
+      input_queue_size: 0
+      output_queue_max: 40
+      output_queue_size: 0
+      queue_strategy: fifo
+      total_output_drop: 0
+    reliability: 255/255
+    rxload: 1/255
+    txload: 1/255
+    type: CSR vNIC
+  GigabitEthernet3:
+    arp_timeout: 04:00:00
+    arp_type: arpa
+    auto_negotiate: true
+    bandwidth: 1000000
+    counters:
+      in_broadcast_pkts: 0
+      in_crc_errors: 0
+      in_errors: 0
+      in_frame: 0
+      in_giants: 0
+      in_ignored: 0
+      in_mac_pause_frames: 0
+      in_multicast_pkts: 0
+      in_no_buffer: 0
+      in_octets: 0
+      in_overrun: 0
+      in_pkts: 0
+      in_runts: 0
+      in_throttles: 0
+      in_watchdog: 0
+      last_clear: never
+      out_babble: 0
+      out_buffer_failure: 0
+      out_buffers_swapped: 0
+      out_collision: 0
+      out_deferred: 0
+      out_errors: 0
+      out_interface_resets: 0
+      out_late_collision: 0
+      out_lost_carrier: 0
+      out_mac_pause_frames: 0
+      out_no_carrier: 0
+      out_octets: 0
+      out_pkts: 0
+      out_underruns: 0
+      out_unknown_protocl_drops: 0
+      rate:
+        in_rate: 0
+        in_rate_pkts: 0
+        load_interval: 300
+        out_rate: 0
+        out_rate_pkts: 0
+    delay: 10
+    description: Network Interface
+    duplex_mode: full
+    enabled: false
+    encapsulations:
+      encapsulation: arpa
+    flow_control:
+      receive: false
+      send: false
+    keepalive: 10
+    last_input: never
+    last_output: never
+    line_protocol: down
+    link_type: auto
+    mac_address: 0050.56bf.7db4
+    media_type: Virtual
+    mtu: 1500
+    oper_status: down
+    output_hang: never
+    phys_address: 0050.56bf.7db4
+    port_channel:
+      port_channel_member: false
+    port_speed: 1000mbps
+    queues:
+      input_queue_drops: 0
+      input_queue_flushes: 0
+      input_queue_max: 375
+      input_queue_size: 0
+      output_queue_max: 40
+      output_queue_size: 0
+      queue_strategy: fifo
+      total_output_drop: 0
+    reliability: 255/255
+    rxload: 1/255
+    txload: 1/255
+    type: CSR vNIC
