+++ Cave_of_Wonders/Cisco/IOS_XE/Show_Interfaces/DevNet_Sandbox_CSR1000v_show_int_mind_map.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<meta http-equiv="X-UA-Compatible" content="ie=edge">
+<title>Markmap</title>
+<style>
+* {
+  margin: 0;
+  padding: 0;
+}
+#mindmap {
+  display: block;
+  width: 100vw;
+  height: 100vh;
+}
+</style>
+<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/markmap-toolbar@0.1.6/dist/style.min.css">
+</head>
+<body>
+<svg id="mindmap"></svg>
+<script src="https://cdn.jsdelivr.net/npm/d3@6.6.0"></script><script src="https://cdn.jsdelivr.net/npm/markmap-view@0.2.6"></script><script src="https://cdn.jsdelivr.net/npm/markmap-toolbar@0.1.6/dist/index.umd.min.js"></script><script>(r => {
+          setTimeout(r);
+        })(() => {
+  const {
+    markmap,
+    mm
+  } = window;
+  const toolbar = new markmap.Toolbar();
+  toolbar.attach(mm);
+  const el = toolbar.render();
+  el.setAttribute('style', 'position:absolute;bottom:20px;right:20px');
+  document.body.append(el);
+})</script><script>((getMarkmap, getOptions, data) => {
+        const {
+          Markmap
+        } = getMarkmap();
+        window.mm = Markmap.create('svg#mindmap', getOptions == null ? void 0 : getOptions(), data);
+      })(() => window.markmap,null,{"t":"heading","d":1,"p":{"lines":[1,2]},"v":"Show Interfaces","c":[{"t":"thead","d":3,"p":{"lines":[2,3]},"v":"","c":[{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Interface"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Description"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Type"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Connected"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Enabled"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Bandwidth"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Port Speed"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Port Duplex"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"IPv4 Address"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Line Protocol"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Operational Status"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"MAC Address"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Media Type"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Delay"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Encapsulation"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Keepalive"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"MTU"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Port Channel Member"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Port Channel Member Interface"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Input CRC Errors"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Input Errors"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Output Errors"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Input Packets"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Output Packets"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"QoS Strategy"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Input Queue Drops"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Total Output Drops"}]},{"t":"tbody","d":3,"p":{"lines":[4,7]},"v":"","c":[{"t":"tr","d":4,"p":{},"v":"","c":[{"t":"td","d":5,"p":{},"v":"GigabitEthernet1"},{"t":"td","d":5,"p":{},"v":"MANAGEMENT INTERFACE - DON'T TOUCH ME"},{"t":"td","d":5,"p":{},"v":"CSR vNIC"},{"t":"td","d":5,"p":{},"v":"Not Connected"},{"t":"td","d":5,"p":{},"v":"True"},{"t":"td","d":5,"p":{},"v":"1000000"},{"t":"td","d":5,"p":{},"v":"1000mbps"},{"t":"td","d":5,"p":{},"v":"full"},{"t":"td","d":5,"p":{},"v":"No IP Address"},{"t":"td","d":5,"p":{},"v":"up"},{"t":"td","d":5,"p":{},"v":"up"},{"t":"td","d":5,"p":{},"v":"0050.56bf.78ac"},{"t":"td","d":5,"p":{},"v":"Virtual"},{"t":"td","d":5,"p":{},"v":"10"},{"t":"td","d":5,"p":{},"v":"arpa"},{"t":"td","d":5,"p":{},"v":"10"},{"t":"td","d":5,"p":{},"v":"1500"},{"t":"td","d":5,"p":{},"v":"False"},{"t":"td","d":5,"p":{},"v":"Stand Alone Port"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"7335"},{"t":"td","d":5,"p":{},"v":"8317"},{"t":"td","d":5,"p":{},"v":"fifo"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"0"}]},{"t":"tr","d":4,"p":{},"v":"","c":[{"t":"td","d":5,"p":{},"v":"GigabitEthernet2"},{"t":"td","d":5,"p":{},"v":"Network Interface"},{"t":"td","d":5,"p":{},"v":"CSR vNIC"},{"t":"td","d":5,"p":{},"v":"Not Connected"},{"t":"td","d":5,"p":{},"v":"False"},{"t":"td","d":5,"p":{},"v":"1000000"},{"t":"td","d":5,"p":{},"v":"1000mbps"},{"t":"td","d":5,"p":{},"v":"full"},{"t":"td","d":5,"p":{},"v":"No IP Address"},{"t":"td","d":5,"p":{},"v":"down"},{"t":"td","d":5,"p":{},"v":"down"},{"t":"td","d":5,"p":{},"v":"0050.56bf.4ea3"},{"t":"td","d":5,"p":{},"v":"Virtual"},{"t":"td","d":5,"p":{},"v":"10"},{"t":"td","d":5,"p":{},"v":"arpa"},{"t":"td","d":5,"p":{},"v":"10"},{"t":"td","d":5,"p":{},"v":"1500"},{"t":"td","d":5,"p":{},"v":"False"},{"t":"td","d":5,"p":{},"v":"Stand Alone Port"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"5"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"fifo"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"0"}]},{"t":"tr","d":4,"p":{},"v":"","c":[{"t":"td","d":5,"p":{},"v":"GigabitEthernet3"},{"t":"td","d":5,"p":{},"v":"Network Interface"},{"t":"td","d":5,"p":{},"v":"CSR vNIC"},{"t":"td","d":5,"p":{},"v":"Not Connected"},{"t":"td","d":5,"p":{},"v":"False"},{"t":"td","d":5,"p":{},"v":"1000000"},{"t":"td","d":5,"p":{},"v":"1000mbps"},{"t":"td","d":5,"p":{},"v":"full"},{"t":"td","d":5,"p":{},"v":"No IP Address"},{"t":"td","d":5,"p":{},"v":"down"},{"t":"td","d":5,"p":{},"v":"down"},{"t":"td","d":5,"p":{},"v":"0050.56bf.7db4"},{"t":"td","d":5,"p":{},"v":"Virtual"},{"t":"td","d":5,"p":{},"v":"10"},{"t":"td","d":5,"p":{},"v":"arpa"},{"t":"td","d":5,"p":{},"v":"10"},{"t":"td","d":5,"p":{},"v":"1500"},{"t":"td","d":5,"p":{},"v":"False"},{"t":"td","d":5,"p":{},"v":"Stand Alone Port"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"fifo"},{"t":"td","d":5,"p":{},"v":"0"},{"t":"td","d":5,"p":{},"v":"0"}]}]}]})</script>
+</body>
+</html>
