+++ Cave_of_Wonders/Cisco/IOS_XE/Show_Inventory/DevNet_Sandbox_CSR1000v_show_inventory.yaml
@@ -1,0 +1,27 @@
+!!python/object/new:genie.conf.base.utils.QDict
+dictitems:
+  main:
+    chassis:
+      CSR1000V:
+        descr: Cisco CSR1000V Chassis
+        name: Chassis
+        pid: CSR1000V
+        sn: 9ESGOBARV9D
+        vid: V00
+  slot:
+    F0:
+      other:
+        CSR1000V:
+          descr: Cisco CSR1000V Embedded Services Processor
+          name: module F0
+          pid: CSR1000V
+          sn: ''
+          vid: ''
+    R0:
+      rp:
+        CSR1000V:
+          descr: Cisco CSR1000V Route Processor
+          name: module R0
+          pid: CSR1000V
+          sn: JAB1303001C
+          vid: V00
