+++ Cave_of_Wonders/Cisco/IOS_XE/Show_Version/DevNet_Sandbox_CSR1000v_show_version.html
@@ -1,0 +1,38 @@
+
+<html><body>
+<h1>Show Version</h1>
+<table style="width:100%">
+  <tr>
+    <th>Hostname</th>
+    <th>Platform</th>
+    <th>Chassis</th>
+    <th>Version</th>
+    <th>Image ID</th>
+    <th>OS</th>   
+    <th>Uptime</th>
+    <th>Last Reload</th>
+    <th>Last Reload Reason</th>
+    <th>Interfaces</th>
+    <th>Total Memory</th>
+    <th>Used Memory</th>
+    <th>Serial Number</th>
+    <th>License Type</th>
+  </tr>
+  <tr>
+    <td>csr1000v-1</td>
+    <td>Virtual XE</td>
+    <td>CSR1000V</td>
+    <td>17.3.1a</td>
+    <td>X86_64_LINUX_IOSD-UNIVERSALK9-M</td>
+    <td>IOS-XE</td>   
+    <td>1 hour, 30 minutes</td>
+    <td></td>
+    <td>reload</td>
+    <td>'Gigabit Ethernet': '3'</td>
+    <td>715705</td>
+    <td>'non-volatile configuration': '32768', 'physical': '3978420'</td>
+    <td>9ESGOBARV9D</td>
+    <td>N/A(Smart License Enabled)</td>
+  </tr>
+</table>
+</body></html>
