+++ Cave_of_Wonders/Cisco/IOS_XE/Show_Version/DevNet_Sandbox_CSR1000v_show_version.json
@@ -1,0 +1,41 @@
+{
+    "version": {
+        "chassis": "CSR1000V",
+        "chassis_sn": "9ESGOBARV9D",
+        "compiled_by": "mcpre",
+        "compiled_date": "Wed 12-Aug-20 00:16",
+        "curr_config_register": "0x2102",
+        "disks": {
+            "bootflash:.": {
+                "disk_size": "6188032",
+                "type_of_disk": "virtual hard disk"
+            }
+        },
+        "hostname": "csr1000v-1",
+        "image_id": "X86_64_LINUX_IOSD-UNIVERSALK9-M",
+        "image_type": "production image",
+        "last_reload_reason": "reload",
+        "license_level": "ax",
+        "license_type": "N/A(Smart License Enabled)",
+        "main_mem": "715705",
+        "mem_size": {
+            "non-volatile configuration": "32768",
+            "physical": "3978420"
+        },
+        "next_reload_license_level": "ax",
+        "number_of_intfs": {
+            "Gigabit Ethernet": "3"
+        },
+        "os": "IOS-XE",
+        "platform": "Virtual XE",
+        "processor_type": "VXE",
+        "returned_to_rom_by": "reload",
+        "rom": "IOS-XE ROMMON",
+        "rtr_type": "CSR1000V",
+        "system_image": "bootflash:packages.conf",
+        "uptime": "1 hour, 30 minutes",
+        "uptime_this_cp": "1 hour, 31 minutes",
+        "version": "17.3.1a",
+        "version_short": "17.3"
+    }
+}
