+++ Cave_of_Wonders/Cisco/IOS_XE/Show_Version/DevNet_Sandbox_CSR1000v_show_version.md
@@ -1,0 +1,5 @@
+
+# Show Version
+| Hostname | Platform | Chassis | Version | Image ID | OS | Uptime | Last Reload | Last Reload Reason | Interfaces | Total Memory | Used Memeory | Serial Number | License Type |
+| -------- | -------- | ------- | ------- | -------- | -- | ------ | ----------- | ------------------ | ---------- | ------------ | ------------ | ------------- | ------------ |
+| csr1000v-1 | Virtual XE | CSR1000V | 17.3.1a | X86_64_LINUX_IOSD-UNIVERSALK9-M | IOS-XE | 1 hour  30 minutes |  | reload | 'Gigabit Ethernet': '3' | 715705 | 'non-volatile configuration': '32768'  'physical': '3978420' | 9ESGOBARV9D | N/A(Smart License Enabled) |
