+++ Cave_of_Wonders/Cisco/IOS_XE/Show_Version/DevNet_Sandbox_CSR1000v_show_version_mind_map.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<meta http-equiv="X-UA-Compatible" content="ie=edge">
+<title>Markmap</title>
+<style>
+* {
+  margin: 0;
+  padding: 0;
+}
+#mindmap {
+  display: block;
+  width: 100vw;
+  height: 100vh;
+}
+</style>
+<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/markmap-toolbar@0.1.6/dist/style.min.css">
+</head>
+<body>
+<svg id="mindmap"></svg>
+<script src="https://cdn.jsdelivr.net/npm/d3@6.6.0"></script><script src="https://cdn.jsdelivr.net/npm/markmap-view@0.2.6"></script><script src="https://cdn.jsdelivr.net/npm/markmap-toolbar@0.1.6/dist/index.umd.min.js"></script><script>(r => {
+          setTimeout(r);
+        })(() => {
+  const {
+    markmap,
+    mm
+  } = window;
+  const toolbar = new markmap.Toolbar();
+  toolbar.attach(mm);
+  const el = toolbar.render();
+  el.setAttribute('style', 'position:absolute;bottom:20px;right:20px');
+  document.body.append(el);
+})</script><script>((getMarkmap, getOptions, data) => {
+        const {
+          Markmap
+        } = getMarkmap();
+        window.mm = Markmap.create('svg#mindmap', getOptions == null ? void 0 : getOptions(), data);
+      })(() => window.markmap,null,{"t":"heading","d":1,"p":{"lines":[1,2]},"v":"Show Version","c":[{"t":"thead","d":3,"p":{"lines":[2,3]},"v":"","c":[{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Hostname"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Platform"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Chassis"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Version"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Image ID"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"OS"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Uptime"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Last Reload"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Last Reload Reason"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Interfaces"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Total Memory"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Used Memeory"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"Serial Number"},{"t":"th","d":5,"p":{"lines":[2,3]},"v":"License Type"}]},{"t":"tbody","d":3,"p":{"lines":[4,5]},"v":"","c":[{"t":"td","d":5,"p":{},"v":"csr1000v-1"},{"t":"td","d":5,"p":{},"v":"Virtual XE"},{"t":"td","d":5,"p":{},"v":"CSR1000V"},{"t":"td","d":5,"p":{},"v":"17.3.1a"},{"t":"td","d":5,"p":{},"v":"X86_64_LINUX_IOSD-UNIVERSALK9-M"},{"t":"td","d":5,"p":{},"v":"IOS-XE"},{"t":"td","d":5,"p":{},"v":"1 hour  30 minutes"},{"t":"td","d":5,"p":{},"v":""},{"t":"td","d":5,"p":{},"v":"reload"},{"t":"td","d":5,"p":{},"v":"'Gigabit Ethernet': '3'"},{"t":"td","d":5,"p":{},"v":"715705"},{"t":"td","d":5,"p":{},"v":"'non-volatile configuration': '32768'  'physical': '3978420'"},{"t":"td","d":5,"p":{},"v":"9ESGOBARV9D"},{"t":"td","d":5,"p":{},"v":"N/A(Smart License Enabled)"}]}]})</script>
+</body>
+</html>
