+++ testbed/testbed.yaml
@@ -1,0 +1,71 @@
+devices:
+  # For IOS-XE devices
+    4500:
+      alias: '4500'
+      type: 'router'
+      os: 'iosxe'
+      platform: cat4500
+      credentials:
+        default:
+          username: {{ your username }}
+          password: {{ your password }}
+      connections:        
+        cli:
+          protocol: ssh
+          ip: {{ your device IP }}
+          arguments:
+            connection_timeout: 360  
+    3850:
+      alias: '3850'
+      type: 'switch'
+      os: 'iosxe'
+      platform: cat3850
+      credentials:
+        default:
+          username: {{ your username }}
+          password: {{ your password }}
+      connections:        
+        cli:
+          protocol: ssh
+          ip: {{ your device IP }}
+          arguments:
+            connection_timeout: 360
+    9300:
+      alias: '9300'
+      type: 'switch'
+      os: 'iosxe'
+      platform: c9300
+      credentials:
+        default:
+          username: {{ your username }}
+          password: {{ your password }}
+      connections:        
+        cli:
+          protocol: ssh
+          ip: {{ your device IP }}
+          arguments:
+            connection_timeout: 360
+  # For JUNOS devices
+    qfx:
+      type: 'switch'
+      os: 'junos'
+      credentials:
+        default:
+          username: {{ your username }}
+          password: {{ your password }}
+      connections:
+        cli:
+          protocol: ssh
+          ip: {{ your device IP }}
+          arguments:
+            connection_timeout: 360
+  #For F5 deivces:
+    F5:
+      alias: 'F5'
+      type: 'F5'
+      os: 'BIGIP'
+      platform: i2600
+      credentials:
+        default:
+          username: {{ your REST API Username }}
+          password: {{ your REST API Password }}
